% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_noise.R
\name{speckle_noise}
\alias{speckle_noise}
\title{Take a numeric and add to it a uniformly distributed random number proportionate to the numeric itself.}
\usage{
speckle_noise(image, variance = 0.04)
}
\arguments{
\item{image}{A numeric. Should be an element of an array.}

\item{variance}{The variance of the uniform distribution the random number is sampled from. The variance is converted into a range by the function.}
}
\description{
Take a numeric and add to it a uniformly distributed random number proportionate to the numeric itself.
}
\examples{
speckle_noise(0.57, 0.1)
speckle_noise(-4)
}
