% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_noise.R
\name{uniform_noise}
\alias{uniform_noise}
\title{Take a numeric and add to it a uniformly distributed random number.}
\usage{
uniform_noise(image, lower_bound = -1, upper_bound = 1)
}
\arguments{
\item{image}{A numeric. Should be an element of an array.}

\item{lower_bound}{The minimum possible value for the random number.}

\item{upper_bound}{The maximum possible value for the random number.}
}
\description{
Take a numeric and add to it a uniformly distributed random number.
}
\examples{
uniform_noise(4.2, 3, 5)
uniform_noise(0.3, -1)
uniform_noise(0.55, upper_bound = 1.2)
uniform_noise(1.4)
}
